--- a/dijkstra/index.js
+++ b/dijkstra/index.js
@@ -1,4 +1,4 @@
 'use strict';
 
-module.exports.graph = require('./lib/graph');
-module.exports.dot = require('./lib/dot');
+exports.graph = require('./lib/graph');
+exports.dot = require('./lib/dot');