--- a/dijkstra/lib/dot.js
+++ b/dijkstra/lib/dot.js
@@ -7,8 +7,8 @@ function graphToDot(graph) {
   output.push('  node [shape="record" style="filled" color="grey79" fillcolor="grey79"];');
 
   graph.forEach((node, i) => {
-    var color = 'grey79';
-    if(Math.random() < 0.5) color = 'red';
+    let color = 'grey79';
+    if (Math.random() < 0.5) color = 'red';
     output.push(`  "${i}" [style="filled" color="grey79" fillcolor="${color}"];`);
     node.forEach((edge) => {
       output.push(`  "${i}" -> "${edge.destination}" [ label="${edge.weight}" ];`);