--- a/dijkstra/lib/dot.js
+++ b/dijkstra/lib/dot.js
@@ -6,11 +6,10 @@ function graphToDot(graph) {
   output.push('digraph dijkstra {');
   output.push('  node [shape="record" style="filled" color="grey79" fillcolor="grey79"];');
 
-  graph.forEach((node, i) => {
+  graph.nodes.forEach((node, i) => {
     let color = 'grey79';
-    if (Math.random() < 0.5) color = 'red';
     output.push(`  "${i}" [style="filled" color="grey79" fillcolor="${color}"];`);
-    node.forEach((edge) => {
+    node.edges.forEach((edge) => {
       output.push(`  "${i}" -> "${edge.destination}" [ label="${edge.weight}" ];`);
     });
   });