--- a/dijkstra/lib/dot.js
+++ b/dijkstra/lib/dot.js
@@ -1,13 +1,13 @@
 'use strict';
 
 function graphToDot(graph) {
-  let output = [];
+  const output = [];
 
   output.push('digraph dijkstra {');
   output.push('  node [shape="record" style="filled" color="grey79" fillcolor="grey79"];');
 
   graph.nodes.forEach((node, i) => {
-    let color = 'grey79';
+    const color = 'grey79';
     output.push(`  "${i}" [style="filled" color="grey79" fillcolor="${color}"];`);
     node.edges.forEach((edge) => {
       output.push(`  "${i}" -> "${edge.destination}" [ label="${edge.weight}" ];`);
@@ -19,4 +19,4 @@ function graphToDot(graph) {
   return output.join('\n');
 }
 
-module.exports.graphToDot = graphToDot;
+exports.graphToDot = graphToDot;