--- a/dijkstra/lib/graph.js
+++ b/dijkstra/lib/graph.js
@@ -2,13 +2,33 @@
 
 const MAX_RANDOM_EDGES = 3;
 
-function getRandomInt(min, max) {
-  return Math.floor(Math.random() * (max - min)) + min;
+class Edge {
+  constructor(destination, weight) {
+    this.weight = weight || getRandomInt(1, 50);
+    this.destination = destination;
+  }
 }
 
-function makeEdge(destination, weight) {
-  weight = weight || getRandomInt(1, 50);
-  return {destination: destination, weight: weight};
+class Node {
+  constructor(edges) {
+    this.edges = edges || [];
+  }
+}
+Node.prototype.push = function (element) {
+  this.edges.push(element);
+};
+
+class Graph {
+  constructor(nodes) {
+    this.nodes = nodes || [];
+  }
+}
+Graph.prototype.push = function (element) {
+  this.nodes.push(element);
+};
+
+function getRandomInt(min, max) {
+  return Math.floor(Math.random() * (max - min)) + min;
 }
 
 function isValidEdge(node, dest, index) {
@@ -16,7 +36,7 @@ function isValidEdge(node, dest, index) {
   if (dest === index) return false;
 
   // avoid duplicate edges
-  for (const edge of node) {
+  for (const edge of node.edges) {
     if (edge.destination === dest) return false;
   }
 
@@ -26,22 +46,23 @@ function isValidEdge(node, dest, index) {
 function generate(numNodes) {
   if (numNodes === 1) throw new Error('Thats not a cool graph. Lets have at least 2 nodes');
 
-  let graph = [];
+  let graph = new Graph();
 
   for (let i = 0; i < numNodes; i++) {
-    let node = [];
+    let node = new Node();
 
     // make circular edges
     //  - assures every node is reachable from every other node
     //  - assures every node has at least one incoming and one outgoing edge
-    node.push(makeEdge(i === numNodes - 1 ? 0 : i + 1));
+    let dest = (i === numNodes - 1 ? 0 : i + 1);
+    node.push(new Edge(dest));
 
     // make random edges
     const numEdges = getRandomInt(0, MAX_RANDOM_EDGES);
     for (let e = 0; e < numEdges; e++) {
-      const dest = getRandomInt(0, numNodes - 1);
+      let dest = getRandomInt(0, numNodes - 1);
       if (isValidEdge(node, dest, i)) {
-        node.push(makeEdge(dest));
+        node.push(new Edge(dest));
       }
     }