--- a/dijkstra/lib/graph.js
+++ b/dijkstra/lib/graph.js
@@ -13,19 +13,21 @@ class Node {
   constructor(edges) {
     this.edges = edges || [];
   }
+
+  push(element) {
+    this.edges.push(element);
+  }
 }
-Node.prototype.push = function (element) {
-  this.edges.push(element);
-};
 
 class Graph {
   constructor(nodes) {
     this.nodes = nodes || [];
   }
+
+  push(element) {
+    this.nodes.push(element);
+  }
 }
-Graph.prototype.push = function (element) {
-  this.nodes.push(element);
-};
 
 function getRandomInt(min, max) {
   return Math.floor(Math.random() * (max - min)) + min;
@@ -46,21 +48,21 @@ function isValidEdge(node, dest, index) {
 function generate(numNodes) {
   if (numNodes === 1) throw new Error('Thats not a cool graph. Lets have at least 2 nodes');
 
-  let graph = new Graph();
+  const graph = new Graph();
 
   for (let i = 0; i < numNodes; i++) {
-    let node = new Node();
+    const node = new Node();
 
     // make circular edges
     //  - assures every node is reachable from every other node
     //  - assures every node has at least one incoming and one outgoing edge
-    let dest = (i === numNodes - 1 ? 0 : i + 1);
+    const dest = (i === numNodes - 1 ? 0 : i + 1);
     node.push(new Edge(dest));
 
     // make random edges
     const numEdges = getRandomInt(0, MAX_RANDOM_EDGES);
     for (let e = 0; e < numEdges; e++) {
-      let dest = getRandomInt(0, numNodes - 1);
+      const dest = getRandomInt(0, numNodes - 1);
       if (isValidEdge(node, dest, i)) {
         node.push(new Edge(dest));
       }
@@ -72,7 +74,7 @@ function generate(numNodes) {
   return graph;
 }
 
-module.exports.generate = generate;
-module.exports.Graph = Graph;
-module.exports.Node = Node;
-module.exports.Edge = Edge;
+exports.generate = generate;
+exports.Graph = Graph;
+exports.Node = Node;
+exports.Edge = Edge;