This is a basic implementation of Dijkstra's algorithm, based on this article https://aos.github.io/2018/02/24/understanding-dijkstras-algorithm/

The test suite suggest this is working. I'd love to get a better understanding though, so next is visualising the algorithm this in the browser.

/cc @mourner