@freenerd 2018-03-22
Turns out the priority comparator wasn't working well, switching it to the right one gave an amazing performance boost https://github.com/freenerd/learning-graph-algorithms/pull/5/commits/995ac082983177944be8e1672ec2d03856a3803e

```
// configuration
let g = graph.generate(100000, 'myseed');
let result = dijkstra.shortestPath(g, 0, 9000);

// old
const queue = new TinyQueue([], (a, b) => a.distance > b.distance);
djikstra: 20488.836ms
djikstra: 20244.876ms
djikstra: 20379.177ms

// new
const queue = new TinyQueue([], (a, b) => a.distance - b.distance);
djikstra: 254.777ms
djikstra: 252.514ms
djikstra: 249.074ms

// other way around has same effect
const queue = new TinyQueue([], (a, b) => b.distance - a.distance);
djikstra: 256.769ms
djikstra: 259.342ms
djikstra: 250.966ms
```